[hw,dma,rtl] Explicitly start and stop hashing
In hardware handshake mode, multiple DMA transfers contribute to the computation
of one digest. To enable the SHA engine  at begin, FW needs to explicitly set
a bit in the control register. The same is true when ending the computation of
the hash. FW needs to set a bit in the control register to stop hashing and to
compute digest.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>