[soc_proxy/rtl] Replace custom sync and filter logic with `prim_filter`
Prior to this commit, the internal wakeup signal was `flop_2sync`ed and
then filtered with custom logic, but it turns out that `prim_filter`
serves this purpose.  The implementation is almost identical, although
the custom filter logic used an AND-reduction, which is biased towards
zero values, while `prim_filter` is not biased.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>