[top_earlgrey,sensor_ctrl] Add tied-off external reset request
As SW is currently not able to support reset request configurations that
are different between Earlgrey and Darjeeling, and we introduced an
additional reset request in Darjeeling in the previous commit, this
commit re-aligns Earlgrey's reset request configuration with
Darjeeling's.  The additional reset request is simply tied off in
Darjeeling.

This commit is **binary-incompatible with Earlgrey-ES**!

Signed-off-by: Andreas Kurth <adk@lowrisc.org>