[mbx] Exclude most CSRs from automated tests
Many of Mailbox's CSRs have side effects or are implemented by a
hardware mechanism that automated CSR tests are not aware of. For this
reason, those CSRs have to be (partially) excluded from automated tests.

To enable progress, at the moment probably too many exclusions are in
place, though. This should be revisited, which is tracked in issue
lowRISC/opentitan-integrated#477 and by the TODO comments in this
commit.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>