[crypto] Simplify the Miller-Rabin primality test.
As of FIPS 186-5, it is permissible to restrict the primes p and q to a
certain value modulo 8. If this value is 3 or 7 (i.e. the candidate
prime is always 3 mod 4), then constant-time Miller-Rabin becomes a lot
simplier. This is nice from a code-complexity and performance
perspective, but the code-size benefit is the real prize. RSA key
generation is close to the IMEM size limit, so this frees up space for
further speed improvements.

Signed-off-by: Jade Philipoom <jadep@google.com>