[top_darjeeling] Remove flash_ctrl
Dummy definitions for flash_ctrl offsets are added as a workaround so that
the flash_ctrl can be removed without having to refactor all dependent tests
and the ROM at once. This needs to be cleaned up and removed. Note that the
SW depending on flash_ctrl is currently not functional.

Signed-off-by: Michael Schaffner <msf@opentitan.org>