[tlgen] Instantiate `socket_1n` for every host
This commit changes `tlgen` so that it instantiates a `socket_1n` for
every host, not just those connected to multiple devices.  Without a
`socket_1n` between a host and a device, there's no hardware that would
detect out-of-bounds addresses in a topology where all hosts are
connected to at most one device.  We did not have this case in current
topologies, but for the `sram` host port of `mbx`, which is only
attached to one device -- this topology will be instantiated in a
subsequent commit -- this is the case.

Cherry-picked from integrated_dev and updated to resolve conflicts.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>