[top_darjeeling] Move JTAG mailbox to separate top-level port
This is required as the JTAG mailbox from the SoC's perspective is in a
different address space than the other mailboxes.  The newly created
debug bus should at some point connect the DMI ports of `rv_dm` and
`lc_ctrl` in the same way that the JTAG mailbox gets connected in this
commit.  (Not doing that right now as the ports of `top_darjeeling` to
which `rv_dm` and `lc_ctrl` are connected are currently frozen.)

Signed-off-by: Andreas Kurth <adk@lowrisc.org>