[top_darjeeling/sw] Add dummy definitions for keymgr
This adds dummy definitions added as a workaround so
that the the old keymgr can be removed without having to
refactor all dependent tests and the ROM at once. These
tests and the DIF need to be refactored so that they
work with keymgr_dpe instead.

Note that the SW depending on the old keymgr is currently
not functional.

Signed-off-by: Michael Schaffner <msf@opentitan.org>