[prim_onehot,cov] Fix coverage when AddrCheck == 0
The prim_onehot_check_if_proxy class can be instantiated when AddrCheck is either on or off. When it is off it should result in no coverage holes for the cp_onehot_addr_fault coverpoint. The previous implementation set option.at_least to zero in this case, which caused a runtime warning. This implementation creates two derived classes from prim_onehot_check_if_proxy, one that is equivalent to the prior class for AddrCheck == 1, and one whose embedded covergroup does not have cp_onehot_addr_fault for AddrCheck == 0.

This removes the option.at_least setting, thus a problematic warning.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>