[pattgen,SiVal] Add features to `pattgen.hjson`
I've split the features of the pattern generator into three groups:
channel, config and complete. The final feature did not really fit in
the other two, so it has its own group. We can use this feature list to
make sure that we test all major features of the block during silicon
validation.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>