[opentitantool] Handle writing of large amounts of UART data
In the `Uart` trait, `read()` and `read_timeout()` may be augmented by
using the `register_nonblocking_read()` function, for `mio`-style
polling for incoming data.  However, the convention is that the
`write()` method is always blocking (there is no
`register_nonblocking_write()`).

The implementation in `transport/common/uart.rs` does tricks with
`port.as_raw_fd()` in order to get a file descriptor out of the
`serialport` library for use with the `mio` library.  It seems that a
side effect of this is that `mio` puts the file descriptor into
non-blocking mode, which may cause `write_all()` to return an `Err()` in
response to getting `EAGAIN` from the operating system due to queues
being full.

I assume that `write_all()` could make multiple OS `write()` calls, and
we have no way of knowing whether the first call succeeded in accepting
some of the bytes, before the second call returned `EAGAIN`.

This CL replaces `write_all()` with a loop attempting individual calls
to `write()`, keeping track of how much has been sent.  Furthermore, in
case of `EAGAIN`, it delays for a few milliseconds before attempting to
transfer the rest of the data, in order to honor the synchronous
contract.

The goal is to be able to cause HyperDebug to continuously transmit at
maximum UART speed, in order to run endurance tests.  5 milliseconds
were chosen because at 115200 baud it amounts to 58 characters, and we
assume that the Linux USB stack has room in its queues for at least a
couple of 64 bytes buffers, meaning that such a delay between write
calls will not likely cause a gap in UART transmission from HyperDebug.
This has been verified with a logic analyzer.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ifedb679a01ec5578738aa2d984c5a45637a4f3b2