[prim_prince,rtl] Split data_state array into two pieces
Recent versions of Verilator complain that data_state is
"multidriven", which turns out to be because the middle
element (data_state[NumRoundsHalf+1]) behaves like it's updated by a
process that's sensitive to different signals from the processes that
update the other elements.

Splitting this array in half (for the forwards and backwards passes,
respectively) seems to evade the warning.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>