[top_darjeeling/dv] Fix activity sampling timing in power virus
Since we removed several IPs from this top-level (and
this test), the timing is a bit different and waiting for 16
cycles to check activity is too much for CSRNG where
the command occasionally completes faster now.

The spurious failures in CI for instance just complete
1 cycle before sampling the activity.

This patch therefore lowers the wait cycles after the
max-power indicator is seen to address that.

Signed-off-by: Michael Schaffner <msf@opentitan.org>