[hw,rv_core_ibex,doc] OTP ROM patch table format changes
The specification defines, describes and document patch match registers
that seem to be Ibex CSRs. None of those exist, neither in Ibex itself
nor in the OT wrapper.

In order to stop relying and documenting non existing registers, this PR
modifies the ROM patch table documentation and defines simpler, CSR
agnostic descriptors for the matching and remapping configuration for
ROM patch regions. This descriptors allow for configuring the existing
Ibex wrapper address translation registers and could also be used for
configuring Ibex CSRs when and if they are implemented.

These descriptors are 2x32 bits words including the base ROM region
address, the patched ROM region size and the SRAM destination base
address. It is up to the patch loader to translate those descriptors
into Ibex wrapper register settings (for now) and potentially Ibex CSR
writes in the future.

Signed-off-by: Samuel Ortiz <sameo@rivosinc.com>