[edn/dv] Add output FIFO errors to edn_err test
Added output FIFO error codes to the err_code_e enum type and to
the err_vseq. This means that now errors for the output FIFO are
tested in the same way as the errors for the other EDN FIFOs.
The H_DataStableWhenValidAndNotReady_A assertion caused the test to
fail for the newly added error codes. The assertion was triggered
because the FIFO clear in the test sequence causes the host data
to be unstable.
Since this is the expected behaviour in this case, the assertion
is disabled shortly before the FIFO clear and reenabled right after.
This commit closes issue #18223.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>