[spi_device/rtl] Fix responses for read/write-only memory
SPI Device's egress SRAM is write-only and its ingress SRAM is
read-only for SW.  This is correctly captured in `spi_device.hjson` but
the RTL does not respond with errors on reads from write-only memory and
on writes to read-only memory.  This causes automated randomized
interconnect tests, which correctly expect these error responses, to
fail.

This commit configures the two `tlul_adapter_sram` that handle TL-UL
accesses to the egress and ingress SRAM to return the appropriate error
responses.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>