[otp] move default fixed secret0 overlay to shared location
Both a manuf test and reference provisioning binary need to make use of
the pre-provisioned secret0 partition overlay. This moves the overlay
to a shared location (in the hw/ip/otp_ctrl/data/ directory) so it can
be reused across build targets.

Additionally this fixes a typo in the data deps of bitstream targets in
manuf tests.

Signed-off-by: Tim Trippel <ttrippel@google.com>