[opentitantool] Do not specify OpenOCD ports
When backend transports choose to start openocd as a sub-process, for
use as part of the implementation of the `Jtag` trait, the caller needs
not care or know which TCP port is being used for communication between
the struct implementing `Jtag` and the sub-process.

This CL instructs openocd to choose any available port, and has the
helper functions in `openocd.rs` recognize the chosen port among the
stderr output of the openocd sub-process, in order to establish
connection.

This will ensure that tests or manual invocations of opentitanlib are
not adversely affected by an unrelated OpenOCD process already having
bound to one of the default ports.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ie7cf04a47f4aae15db9a8e6d90266c879dd5b2a9