[flash_ctrl,dv] update expected double bit error set
When multiple writes are issued to the same address of flash device,
data corruption is exected and detected by checking ecc.
However, flash device can only be update from 1 to 0 (not the other way),
multiple writes can make data bits converge to 0 and this can cause
to generate double bit error.
So expected double bit error in the error tests, is updated such that
it can only be set only after tb caculates ecc and confirms it to generate double bit error.

Signed-off-by: Jaedon Kim <jdonjdon@opentitan.org>