[opentitantool] More flexible argument types for OpenOCD command
Rather than accumulate all arguments in a `Vec<&str>`, this CL makes it
such that arguments are given one at a time to `Command`, as any `S:
AsRef<OsStr>`.  This paves the way for passing `PathBuf` directly, even
if it contains non-utf8 characters.

Also, the debug printing of the `cmd` object itself, rather than
separately printing the program and its arguements, results in the
command arguments being individually quoted, such that a developer could
copy-paste the entire invocation into a shell, for debugging.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Iee661f2f78dd7924499a983130581b00f37c0dc7