[opentitantool] Avoid relying on udev for Sam3X serial ports
Chrome OS has a "chroot" development environment and OpenTitan CI uses
VMs.  The HyperDebug backend has no trouble with accessing UARTs inside
chroot or VMs, but the CW310 backend does not automatically recognize
its UARTs.  This is because udev enumeration does not work properly in
these two kinds of sandboxes.

There is one major reason why the HyperDebug backend does not use
`serialport::available_ports()` (other than the fact that I did not know
about the method).  The HyperDebug family of debuggers (consisting of
HyperDebug, a range of Google debuggers: "Servo Micro", "C2D2", as well
as Titan chips running GSC firmware) present a varying number of UARTs
as USB interfaces, and they label each such interface with a text string
in the USB descriptor.  We want that text string to become the "name" of
the UART for `Transport::uart()`, rather than merely numbering the
UARTs, as their ordering may vary between the various debuggers listed
above.  It seems that `serialport::available_ports()` only advertises
the USB serial number of the debugger that each UART belongs to, not any
identification of the particular USB interface, and so it is
insufficient for the HyperDebug backend transport.

Since it is now apparent that the HyperDebug way of discovering the like
between USB devices, their USB interfaces, and logical `/dev/ttyXXX`
devices actually work in chroot and VM environments, unlike the
udev-based `serialport::available_ports()`, I thought that it would make
sense to make both CW310 and HyperDebug backend use the same approach,
and I thought that I could solve my own problems with using CW310 inside
chroot by changing CW310.  At the same time, if this change can
eliminate the need for passing UART parameters in CI, that would be an
added bonus.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I9ff872e96edcdddcf02f2e492eade0de8812ea1b