[sw] Add support for Zephyr's devicetree APIs
Zephyr has some very useful work that transforms a devicetree into a
single C header that represents a large amount of the information
present in the source. The generated header can be queried purely with
macros (Zephyr's devicetree APIs), and with these features and some
modifications, many OpenTitan tests and testutils may be written to
target an abstract top.

Add support for Zephyr's devicetree APIs. Import the repository, and
patch the tree for the following purposes:
  - Add the ability to query the interrupt-parent from a given IRQ.
  - Use the IRQ number directly from the device tree in lookups, instead
    of Zephyr's special multi-level encoding, which obscures the ID at
    the PLIC.
  - Reduce the deps surface for the devicetree APIs so only macros are
    #include'd. devicetree.h didn't need more than sys/util_macro.h.

Add bazel rules to process device trees in two steps:
  1. Gather all device tree sources and use the C preprocessor to
     resolve includes and defines and consolidate the device tree into a
     single file.
  2. Run the processed source and bindings through Zephyr's
     gen_defines.py script, which creates the generated header.

The output of the bazel rule is a header target that can be used in
downstream cc_* targets. The intention is to make this header part of
the standard includes for a particular platform or execution
environment.

Signed-off-by: Alexander Williams <awill@opentitan.org>