[opentitantool] Eliminate POR pin from board.rs
Checking the version of the currently running bitstream relies only on
console UART and reset GPIO pin, and need not be coded at the low level
of the transport backend implementation.

This PR moves the logic up above the TransportWrapper, allowing the use
of aliases from JSON configuration files, rather than Rust traits.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I107edc48e309b86b2f902211daf00f3efb4136cb