[sw,device,rom] Initial Darjeeling base ROM
The base ROM is a minimal ROM that aims at jumping to the second stage
ROM as early as possible.

With that in mind, the base ROM configures the necessary IP blocks, sets
as many ePMP as possible, set the C runtime up and then jumps to the
second stage ROM.

The base ROM must load, very and enable patches for the second stage
ROM, although this patch does not support that feature yet.

Signed-off-by: Loïc Lefort <loic@rivosinc.com>
Signed-off-by: Samuel Ortiz <sameo@rivosinc.com>