[chip-test,sival] Extend OTBN randomness app
This commit extends the OTBN randomness app. It now accepts an
argument telling the app how many additional words of entropy should
be consumed. The consumed entropy only comes from the RND register,
since the randomness from the URND register comes from a PRNG. This
serves as the basis for the edn_auto_mode sival test.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>