[sw/crypto] Add P-384 ECDH binary entrypoint
This commit
- adds OTBN binary code for P384 ECDH entrypoint. To be able to have
keypair generation and shared key computation in one binary projective
add in p384_base.s is optimized by removing redundant instructions. This
frees enough IMEM to have a single ECDH entrypoint file.
- adds a corresponding test for ECDH shared key generation.
- adds OTBN binary code to check if a given point is a valid P-384 curve
point. For ECDH, to check if a given public key is valid, this binary
has to be called in advance of generating a shared key. Unfortunately,
due to IMEM limitations, this can't be a part of the ECDH entrypoint and
therefore has to be handled separately. It is important for later C Ibex
driver development to consider this.
- adds a corresponding test for curve point validation.

Signed-off-by: Moritz Wettermann <moritz.wettermann@gi-de.com>