[ottf] Allow the status library to call a custom abort handler
The status library should not call abort() directly on most cases.
Indeed, the only case where this library is used outside without
ottf_main is in the test_rom. With this commit, ottf_main can be
notified when the test reports a success/failure and print a stack
trace or other information.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>