[ottf] Add a utils library to handle host synchronization
Many dv tests rely on host backdoor writes to synchronization. This
usually takes the form of a global volatile variable and active
spinning while waiting for the DV environment to do an option:
  IBEX_SPIN_FOR(kPhase != kPhase3, kTestTimeout);
This operation cannot be supported in this form in sival where
the host cannot access the memory directly. This commit introduces
a new macro OTTF_WAIT_UNTIL that behaves exactly like
IBEX_SPIN_FOR in DV, but waits for ujson command on real devices.
The macro checks the condition after every command and returns
when it is true.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>