[doc/security] Publish reset vs. non-reset flops rationale
In the past, there have been many discussions on the pros and cons of
reset vs. non-reset flops for registering security sensitive assets in
OpenTitan. While these discussion concluded quite some time ago already,
we've never managed to transfer the summary into MarkDown and publish
it as part of our documentation. This is what this commit changes.

Other key authors of this document include:
- Tim Chen - @tjaychen
- Ziv Hershman - @zi-v
- Chris Gori - @cdgori
- Scott Johnson - @sjgitty
- Johann Heyszl - @johannheyszl

This resolves lowRISC/OpenTitan#2603.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>