[dv] Add original format string to sw_logger_if
Save the original format string and make it available for conditional
expressions for waits. Without this, it is quite difficult to wait for a
specific form of a string and get the correct printed_arg values. The
printed_log already has the substitutions, but the point of printed_arg
is to get values that are _not_ already known.

Signed-off-by: Alexander Williams <awill@opentitan.org>