[boot_svc] Do not use rnd in the empty message
The empty message cannot be initialized with random data using the
silicon_creator `rnd` driver.  That driver reads OTP from SW_CREATOR_CFG
and that partition is locked by the ROM_EXT.  Attempting to use the
`rnd` driver from the owner stage will result in a cpu fault.

Since the point of the empty message is to overwrite any existing
content in the boot_svc retention RAM, writing zeros is sufficient.

1. Rename the `rand_data` field to `payload`.
2. Do not initialize the payload with random data.

Signed-off-by: Chris Frantz <cfrantz@google.com>