[opentitantool] Add gpio MultiSet to proxy protocol
The gpio trait has a `set()` method that allows simultaneously setting
mode, logic level and weak pull mode.  This both improves performance on
HyperDebug (by saving roundtrips of the textual protocol), and
eliminates glitches.

For transports that do not natively support this, the trait does provide
a default implementation which resolves to two or three method calls.

It appears that when this was added, I failed to extend the proxy
protocol to be able to pass the `set()` method natively, and instead,
the operation is currently broking into multiple sub-operations on the
client side, which is inefficient.  This CL adds straigt-forward
code to pass the multi-set operation via the proxy protocol.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I77c8457279096afd6ed43192c234ebe92c90f89b