[spi_device] Fix up some address / size info
The hjson, pkg, and DV files were out of sync for the SRAM breakdown.
Have spi_device_env_pkg take its information from the RTL, so there is a
single source of truth. Also fix up the hjson to reflect the true
minimum size of the ingress memory.

Adjust the breakdown in spi_device_pkg to make the payload buffer index
have the correct value.

Signed-off-by: Alexander Williams <awill@opentitan.org>