[hw,dma,rtl] Clear STATUS. when starting a new transfer
Previously, the FW needs to manually clear the done signal after a transfer,
to allow a polling loop to get the fresh data. This PR changes the
hardware to automatically clear the done bit on new transfer. This
removes the requirement on the FW to clear the DONE bit after a
transfer. Also clear the error code and SHA2 digest/valid.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>

[hw,dma,rtl] Clear status bits on new transfer

Clear the error and error code. Clear the abort. Use the same
meachnaism to clear the sha2_digest_valid and the digest.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>