[top, fpga] Rework SCA/FI trigger signal selection and generation
This commit modifies the SCA/FI trigger signal generation on FPGA to
also add the option for a fully software-controlled trigger. This is
required to prepare the SCA capture setup for the penetration testing
of the ES silicon.

To control the trigger type dynamically, this commit further adds a new
simple serial command 't'.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>