[opentitanlib] Refactor OpenOCD handling
This commit separates the logic interfacing with the OpenOCD itself with
the logic that drives the upper-level Tcl command. This allows the
lower-level OpenOCD server handler to be reused when not interacting at
TAP level.

A notable change made in this commit is that we eliminate the use of
OpenOCD command completely except for port configuration. `noinit` is
used to launch OpenOCD in configuration mode and all other configuration
is done via the Tcl interface. This allows much code simplification.

`OpenOcdServer` is renamed to `OpenOcdJtagChain` to represent what it
is.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>