[opentitanlib] use buffered stream for OpenOCD handler
Currently the code requries one syscall for each byte read, which is
very inefficient. Wrap the BufReader to the TcpStream to avoid this.

It also turns out that the way we read previously doesn't capture
end-of-stream until the next command. This has been fixed and the
e2e_openocd_debug_test is changed accordingly.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>