[opentitanlib] Remove interior mutability for `OpenOcdJtagChain`
Currently `OpenOcdJtagChain` uses interior mutability everywhere where
it could just be using mutable references. This commit does that
conversion. In order to do so, `Rc<dyn Jtag>` is being converted to
`Box<dyn Jtag>`.

I also take the opportunity to add lifetimes to `dyn Jtag` so it won't
outlive `TransportWrapper`, although it is not necessary currently
because `OpenOcd` is `'static`.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>