[opentitantool] Declare trait methods for I2C device
The HyperDebug debugger is able to act as I2C device as well as I2C host.

This CL attempts to add methods to the `i2c::Bus` trait to allow
opentitantool to ask to put a given I2C instance into "device mode", and
then control how the debugger should behave, when addressed by the I2C
host (which would be the OpenTitan chip).

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: If303d246b464a5988e04bb94fc6196206f0d3096