[rom/e2e] Fix shutdown watchdog
The problem turned out to be really stupid: LOG_INFO prefixes
every message with I<number> where <number> has 6 digits. For
some reason, the BUILD specified one zero too many so the harness
failed to match the message.
During debugging, I added a function to regularly print a status
which is quite useful to help narrow down problems and might be
useful in CI as well so I decided to leave it.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>