[keymgr_dpe, dv] modify current_internal_key array in scb
modifies the current_internal_key array in the scoreboard to
be an array of DpeNumKeySlots entries. Each entry is a struct of
key_shares_t and keymgr_dpe_key_policy_t. Then replaces all references
to the current_internal_key in the scb and calls to the keymgr_dpe_if.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>