[keymgr_dpe, dv] modify process_kmac_data_rsp()
modifies the behavior of the process_kmac_data_rsp() task
of the keymgr_dpe scb. Adds some debug prints, and clarifying
comments. Removes an update_kdf_key() call. This is because, during
the kmac_data_rsp, the kmac_o key could be 0, since it's only visible
on the output port of keymgr_dpe when a kmac_data_req is made.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>