[ci] Better bitstream build cache filtering
We want to avoid rebuilding bitstreams if possible. To do this we
look at the list of changed files and determine if a rebuild is
warranted.

Currently the filtering is entirely hardcoded rules. This commit
uses Bazel to get a list of dependencies required for the bitstream,
and filter the changes based on that list.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>