[sca, aes] Increase max number of segments per batch for fvsr captures
The number of encryptions per batch (number of segments) is determined
by the ChipWhisperer Husky scope and depends on the number of samples
per segment. Now that we clock ratios changed and we're capturing
fewer samples per segment, the number of segments increases and the
buffers storing the pre-generated plaintexts and keys need to be
increased. Previously, the buffer size was chosen very pessimistically.
It should be perfectly fine to allocate space for up to 256 plaintexts
and keys (8 kBytes) as we're having 128 kBytes of main SRAM on all
relevant top levels.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>