[keymgr_dpe, dv] modify process_tl_access "start"
modifies the behavior of the "start" csr access in
process_tl_access() for the keymgr_dpe scb.
When in StWorkDpeReset state, latches the OTP when op is
DpeOpAdvance. Also further checks and updates the kdf key
in the keymgr_dpe_if for comparison with the keymgr_dpe.kmac_key_o
key that is sent to the kmac engine.

adds some debug prints in a few functions.

adds a get_key_slots function to receive the current
active key slot.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>